refactor: remove babel transpilation
BREAKING CHANGE:

With the source code migrated to ES6, usage in client tests with testemJS & PhantomJS or other ES5 browser are no longer supported.
