--- .babelrc
@@ -1,4 +1,0 @@
-{
-  "presets": ["env"],
-  "plugins": ["transform-object-rest-spread"]
-}
